requirements.txt
# FastAPI and server
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
psycopg[binary]>=3.1.18
alembic>=1.13.1

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.2

# Kafka
aiokafka>=0.10.0

# AI - Gemini
google-generativeai>=0.3.2

# Validation and utilities
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# HTTP client for external APIs
httpx>=0.26.0

# Async support
asyncpg>=0.29.0
